requirements changed: sub commands now to be stateless
